feat: add generators for DbSets and DI
The [GenerateDbSets] attribute can be applied to the project's DbContext implementation to automatically generate the DbSet properties in a partial class by the same name.

The [RegisterService] attribute can be applied to any class to automatically add it to an extension method named for the assembly the class comes from. Constructors will also be generated automatically based on the unassigned private readonly fields in the class.