fix: output Anima.Generators as analyzer
Updates the csproj so that the DLL is output into the analyzers folder in the nupkg and not output as a standard library. This was needed so that the package will be installed as an analyser instead of a library by consuming projects.

Also resolved some dependency issues where Microsoft.CodeAnalysis packages that were used for local development were also required as dependencies on the live packages. This is no longer the case.

Removed Version tag from Anima.Generators.csproj as well so it will now correctly use the auto-version from Directory.Build.props as it was meant to.