fix: internalize Pluralize.NET code
The previous fix of copying the DLL to the analyzers directory failed because the DLL isn't actually output when the source generator libraries are compiled. The Pluralize.NET source and license has instead been added to the generator utilities project to avoid issues with trying to reference the NuGet package.