feat: improve source generation + extension methods
RegisterService source generator has been updated to allow for marking abstract base classes with an attribute to prompt source generation of a base constructor to populate dependency fields.

Some test classes and sample projects have been added for the source generators so they can be worked with more easily.

The "Anima" project has been added for generally useful classes and methods that are likely to be used in most projects.

Solution has been configured for packaging locally, since pushing constant broken releases to NuGet for testing was not ideal.